Store number of rows directly
Instead of storing the number of rows in an array with only one element it can
directly store the number itself. It's also using the confusing prefix
`$empty_` even though it does not contain whether it's empty but the number of
lines. (It's actually the other way around, it'll be 0 (or false) when it's
empty)

This also provides a function to get number of rows when using `COUNT()` in the
SQL query.