Remove usage of mysql_real_escape_string
Calling `mysql_real_escape_string` is not necessary when that result is fed
into the info box string because it does not handle HTML entities but MYSQL
query specific entities.

This fixes an issue with the PDO implementation in 13ae949 as calling the
function doesn't work if the file is not connected to the database.