Separate comments and photo markers
With 17d93d8 the markers having a comment or photo use a different (yellow)
marker. And then 086eb83 changed that into a purple marker. As it is intended
that in the future the marker uses a different marker having a photo anyway and
the yellow marker is still available it can be used as a place holder for now.