Replace mysql specific calls with PDO usage
To prevent SQL injections, it is using PDO which supports binding parameters.
Also the `mysql_query` function has been deprecated in favour of PDO or
`mysqli`. Using PDO allows in theory to use different databases in the future.

To support different databases the connection must be more generically
defined which is prepared already in this patch but doesn't require to change
the configuration file yet.