Update Google Maps JavaScript API to version 3
This updates the Google Maps JavaScript API from version 2 to version 3. With
future APIs the overview map will be removed so the corresponding setting is
deprecated as well. Also the map types have changed but the old values will be
mapped. Because any notice would be printed at the top and thus generate
invalid HTML, it does not yet inform the user about the changes.

Please note that version 3 needs a new API key generated for that version!

This also only adds the center cross related code if this option is actually
enabled.

Instead of building the polyline once at the end, it is now assigning a
polyline to the trip where adding a marker will cause it to be added to the
polyline automatically.