Fix zoom level when using live tracking
The zoom level couldn't be calculated when using live tracking with 07d5546.
Before that patch, when using live tracking, it determined the zoom level for
the bounds and then subtracted the zoom level defined. But the zoom level will
be always maximum as it is only one point, so it can always subtract the
defined zoom level from 16.