Use same conversion factor
Always use the same conversion factor to convert between miles and kilometers
so that the total distance in the last marker and the summary are the same.
This was introduced in 4fc6002c.

This also fixes the issue that the first marker shows the total distance from
the previous calculation introduced in fcd60dbf for the metric variant and in
4fc6002c also for the imperial unit.