Lowercase kilometer abbreviation
According to the SI definition, the abbreviations of units and prefixes are
standardized and the `k` in `kilo` is to be written in lower case.

While it technically standardized not all languages actually follow that rule
so this patch is only changes the languages in which I am sure that the
lowercase variant is usually used.