Select English entry by default
If there is no language selected it will default to English but actually
preselect the first loaded language (usually the first alphebetically). This
was because the English language doesn't define the language in English as it
is the same as in its language.

But with ce41a7c the display used the English name to associate the used
language with the preselected one.