Analyze found trips for a user first
This is restructuring how it creates the trip selection box, so that it now
properly selects either the `None` option or the first non-generic option if
the trip specified is invalid.

Previously it only knew that no trip matched after it already had added them
all to the toolbox and so couldn't select another one. This lead to the it
selecting `None` implicitly even though it could select the first actual trip.