Properly unset ID if trip ID is invalid
With 3ae4b9f5 the ID could be implied if the trip ID is used. But when the
trip ID is invalid, it instead returns `false` which looks to later
comparisons that the ID is actually set. But it should default to the first
user if the site is public.