Create balloon text in JS function
Instead of creating the balloon text in a PHP code it adds a JS function which
creates the HTML from it. This also is based on kilometers and calculates the
imperial unit inside the function.

As a side effect it can now determine the icon in `Trip.appendMarker` almost
from the data given. The only exception is a custom marker.