Convert download script to use custom PDO class
The download script used deprecated `mysql`-PHP calls while a PDO replacement
has been introduced with e3ac335 and this is applying the replacement to the
script.

Because not all database drivers support counting the rows of a `SELECT`
statement, these have been replaced by `fetchAll()` and simply counting the
result.