Support translations with placeholders in JS
The JavaScript translator now supports placeholder in the style of "{NUMBER}"
to support strings like "X of Y" which is now "{0} of {1}". So this is now
also translating the label which shows the current point number and number of
points.

The reason it is not using simply a translation of "of" is that there might be
cases in which other languages format the string differently and placeholders
are more flexible. If there is a language which swaps for example the current
and maximum, it is possible by doing something like "{1} foo {0}".