Simplify Google's cellid lookup
Instead of constructing the lookup code manually, it can use PHP's `pack`
function directly. This also prevents the issue that when the separate
parameters are used (`mcc`, `mnc`, `cid` and `lac`) it did not convert them to
a hexadecimal string.

For reading the answer it won't assume that the machine is using little endian
but instead convert the value so that it's interpreted as a big endian.