Calculate distance and time in JavaScript
When loading part of the trip it's not possible to know the total time elapsed
and to know the total distance so let the JavaScript marker figure it out.

This also renames the `totalDistance` property of each marker's data to
`distanceToHere` as it is more descriptive. It attaches the data to the marker
too, so it is possible to access it later.