Fix account creation and detecting missing ones
It previously couldn't detect that an account didn't exist, so it never
returned that result. This is because fetching a row returns `false` and not
`null`. Additionally it couldn't properly create a new account as it overwrote
the password variable with the hash, and when logged in later it obviously
didn't work because it hashed the hash again and they didn't match.