Add test suite
This adds a few tests for several files including the requests script. It
enables the automated usage on Travis. At the moment it doesn't initalize a
custom database for the tests, so running it in a production environment is
not recommended.

To actually test the requests script, it would need to be required multiple
times. So instead it has moved the handling inside a function which can be
called instead. Additionally that script now doesn't `die()` for each tested
case, as it'd kill the test. Only a the code paths are actually changed which
are tested, others still calling `die()`.

It is using `expectException` which isn't available on Travis (yet?), so it'll
ignore those tests if it is not present.